feat(seo-ux): camp landings, FAQ schema, WhatsApp, and enrollment JSON-LD
Improve camp SEO pages, open FAQs by default site-wide, add WhatsApp CTA,
and align JSON-LD with visible content on STEAM, HS math, and enroll routes.
