+++ next.config.ts
@@ -51,6 +51,18 @@ const nextConfig: NextConfig = {
         protocol: 'https',
         hostname: 'images.unsplash.com',
       },
+      {
+        protocol: 'https',
+        hostname: 'patch.com',
+      },
+      {
+        protocol: 'https',
+        hostname: 'assets.activityhero.com',
+      },
+      {
+        protocol: 'https',
+        hostname: 'www.6crickets.com',
+      },
     ],
     formats: ['image/avif', 'image/webp'], // Use modern image formats
     deviceSizes: [640, 750, 828, 1080, 1200, 1920, 2048, 3840],
