+++ src/app/[locale]/courses/high-school-math/layout.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next'
 import FAQSchema from '@/components/schema/FAQSchema'
-import { HIGH_SCHOOL_MATH_FAQ_JSONLD } from '@/lib/schema/course-hub-jsonld-faqs'
+import { HS_MATH_VISIBLE_FAQS } from '@/lib/schema/course-hub-jsonld-faqs'
 import { generateMetadataFromPath } from '@/lib/seo/metadata'
 import { generateCourseSchema, generateBreadcrumbSchema } from '@/lib/seo/structuredData'
 import { absoluteSiteUrl } from '@/lib/publicPath'
@@ -67,7 +67,7 @@ export default async function HighSchoolMathLayout({
         type="application/ld+json"
         dangerouslySetInnerHTML={{ __html: JSON.stringify(breadcrumbSchema) }}
       />
-      <FAQSchema faqs={HIGH_SCHOOL_MATH_FAQ_JSONLD} />
+      <FAQSchema faqs={HS_MATH_VISIBLE_FAQS} />
       {children}
     </>
   )
