+++ src/components/camps/ScheduleSection.tsx
@@ -6,15 +6,23 @@ type ScheduleSectionProps = {
   page: CampLandingPage;
 };
 
+const DEFAULT_SCHEDULE_INTRO =
+  "Specific dates and daily hours are confirmed during enrollment. What stays consistent is structure: preparation, guided work, and review.";
+
 export function ScheduleSection({ page }: ScheduleSectionProps) {
+  const scheduleIntro =
+    page.scheduleIntro === null
+      ? null
+      : page.scheduleIntro ?? DEFAULT_SCHEDULE_INTRO;
+
   return (
     <SectionContainer id="schedule" className="bg-white border-t border-slate-100">
       <div className="max-w-5xl">
         <p className="text-sm font-semibold uppercase tracking-wide text-[#1F396D]">Schedule & enrollment flow</p>
         <h2 className="mt-2 text-2xl sm:text-3xl font-semibold text-slate-900">How the week is organized</h2>
-        <p className="mt-3 max-w-2xl text-slate-600">
-          Specific dates and daily hours are confirmed during enrollment. What stays consistent is structure: preparation, guided work, and review.
-        </p>
+        {scheduleIntro ? (
+          <p className="mt-3 max-w-2xl text-slate-600">{scheduleIntro}</p>
+        ) : null}
 
         <div className="mt-8 overflow-x-auto rounded-xl ring-1 ring-slate-200">
           <table className="min-w-full border-collapse text-left text-sm">
