+++ src/components/layout/WhatsAppFloatingButton.tsx
@@ -1,0 +1,37 @@
+'use client';
+
+import { usePathname } from 'next/navigation';
+import { useChatbot } from '@/contexts/ChatbotContext';
+import { WHATSAPP_CHAT } from '@/lib/whatsapp';
+
+export function WhatsAppFloatingButton() {
+  const pathname = usePathname() ?? '/';
+  const { isOpen } = useChatbot();
+
+  if (isOpen) {
+    return null;
+  }
+
+  return (
+    <div className="group fixed bottom-6 right-6 z-[105]">
+      <span
+        className="pointer-events-none absolute right-full top-1/2 mr-2 -translate-y-1/2 whitespace-nowrap rounded-full bg-gray-900 px-3 py-1.5 text-xs text-white opacity-0 transition-opacity group-hover:opacity-100 group-focus-within:opacity-100"
+        role="tooltip"
+      >
+        {WHATSAPP_CHAT.tooltip}
+      </span>
+      <a
+        href={WHATSAPP_CHAT.getHref(pathname)}
+        target="_blank"
+        rel="noopener noreferrer"
+        aria-label={WHATSAPP_CHAT.ariaLabel}
+        className="flex h-[52px] w-[52px] items-center justify-center rounded-full bg-[#25D366] shadow-lg transition-transform hover:scale-105 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-[#25D366] focus-visible:ring-offset-2"
+      >
+        <svg viewBox="0 0 24 24" fill="white" width="28" height="28" aria-hidden>
+          <path d="M17.472 14.382c-.297-.149-1.758-.867-2.03-.967-.273-.099-.471-.148-.67.15-.197.297-.767.966-.94 1.164-.173.199-.347.223-.644.075-.297-.15-1.255-.463-2.39-1.475-.883-.788-1.48-1.761-1.653-2.059-.173-.297-.018-.458.13-.606.134-.133.298-.347.446-.52.149-.174.198-.298.298-.497.099-.198.05-.371-.025-.52-.075-.149-.669-1.612-.916-2.207-.242-.579-.487-.5-.669-.51-.173-.008-.371-.01-.57-.01-.198 0-.52.074-.792.372-.272.297-1.04 1.016-1.04 2.479 0 1.462 1.065 2.875 1.213 3.074.149.198 2.096 3.2 5.077 4.487.709.306 1.262.489 1.694.625.712.227 1.36.195 1.871.118.571-.085 1.758-.719 2.006-1.413.248-.694.248-1.289.173-1.413-.074-.124-.272-.198-.57-.347z" />
+          <path d="M12 0C5.373 0 0 5.373 0 12c0 2.123.554 4.118 1.528 5.855L.057 23.03a1 1 0 001.224 1.224l5.243-1.462A11.945 11.945 0 0012 24c6.627 0 12-5.373 12-12S18.627 0 12 0zm0 22c-1.885 0-3.65-.52-5.16-1.426l-.36-.214-3.733 1.041 1.05-3.645-.235-.374A9.96 9.96 0 012 12C2 6.477 6.477 2 12 2s10 5.477 10 10-4.477 10-10 10z" />
+        </svg>
+      </a>
+    </div>
+  );
+}
