+++ src/components/math-finals-practice/MathFinalsPracticeLanding.tsx
@@ -16,6 +16,7 @@ import {
   AccordionItem,
   AccordionTrigger,
 } from '@/components/ui/accordion'
+import { getDefaultOpenFaqValues } from '@/lib/faq-accordion'
 import FormPrivacyConsent from '@/components/form/FormPrivacyConsent'
 import { MATH_FINALS_PRACTICE_FAQS } from '@/data/math-finals-practice-faqs'
 import {
@@ -621,7 +622,11 @@ export function MathFinalsPracticeLanding() {
               long-term tutoring plan and it is not designed to reteach an entire course in a short sequence.
             </p>
           </div>
-          <Accordion type="single" collapsible className="space-y-4">
+          <Accordion
+            type="multiple"
+            className="space-y-4"
+            defaultValue={getDefaultOpenFaqValues(MATH_FINALS_PRACTICE_FAQS.length, (i) => `q-${i}`)}
+          >
             {MATH_FINALS_PRACTICE_FAQS.map((faq, i) => (
               <AccordionItem
                 value={`q-${i}`}
