+++ src/components/seo/CourseFAQ.tsx
@@ -6,6 +6,7 @@
 'use client'
 
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from '@/components/ui/accordion'
+import { getDefaultOpenFaqValues } from '@/lib/faq-accordion'
 import { HelpCircle } from 'lucide-react'
 import { StructuredDataScript } from './StructuredDataScript'
 import { generateFAQPageSchema } from '@/lib/seo/structuredData'
@@ -53,7 +54,11 @@ export function CourseFAQ({
           )}
         </div>
 
-        <Accordion type="single" collapsible className="space-y-4">
+        <Accordion
+          type="multiple"
+          className="space-y-4"
+          defaultValue={getDefaultOpenFaqValues(faqs.length, (index) => `item-${index}`)}
+        >
           {faqs.map((faq, index) => (
             <AccordionItem 
               key={index}
